04630: moved test ino AddressAliasIdFuzzing and removed redundant classes
