+++ hedera-node/test-clients/src/main/resource/eet-config/id-fuzzing.properties
@@ -7,3 +7,4 @@ randomAccount.bias=60
 randomAccount.ceilingNum=100
 randomAccountUpdate.bias=30
 randomTransfer.bias=100
+randomEthereumTransactionTransfer.bias=100
